remove unused resources + modify test files
