+++ datadog/tests/data_source_datadog_csm_threats_multi_policy_agent_rules_test.go
@@ -7,14 +7,14 @@ import (
 	"testing"
 
 	"github.com/DataDog/datadog-api-client-go/v2/api/datadogV2"
-
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 
 	"github.com/terraform-providers/terraform-provider-datadog/datadog/fwprovider"
 )
 
-func TestAccCSMThreatsMultiPolicyAgentRuleDataSource(t *testing.T) {
+func TestAccCSMThreatsMultiPolicyAgentRulesDataSource(t *testing.T) {
+	t.Parallel()
 	ctx, providers, accProviders := testAccFrameworkMuxProviders(context.Background(), t)
 
 	policyName := uniqueAgentRuleName(ctx)
@@ -26,23 +26,25 @@ func TestAccCSMThreatsMultiPolicyAgentRuleDataSource(t *testing.T) {
 		tags              = ["host_name:test_host"]
 	}
 	`, policyName)
+
 	agentRuleName := uniqueAgentRuleName(ctx)
 	agentRuleConfig := fmt.Sprintf(`
 		%s
 		resource "datadog_csm_threats_multi_policy_agent_rule" "agent_rule_for_data_source_test" {
 			name              = "%s"
-            policy_id         = datadog_csm_threats_policy.policy_for_test.id
+			policy_id         = datadog_csm_threats_policy.policy_for_test.id
 			enabled           = true
 			description       = "im a rule"
 			expression 		  = "open.file.name == \"etc/shadow/password\""
+			product_tags      = ["compliance_framework:PCI-DSS"]
 		}
 	`, policyConfig, agentRuleName)
-	dataSourceName := "data.datadog_csm_threats_agent_rules.my_data_source"
+	dataSourceName := "data.datadog_csm_threats_multi_policy_agent_rules.my_data_source"
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV5ProviderFactories: accProviders,
-		CheckDestroy:             testAccCheckCSMThreatsAgentRuleDestroy(providers.frameworkProvider),
+		CheckDestroy:             testAccCheckCSMThreatsMultiPolicyAgentRuleDestroy(providers.frameworkProvider),
 		Steps: []resource.TestStep{
 			{
 				// Create an agent rule to have at least one
@@ -52,7 +54,7 @@ func TestAccCSMThreatsMultiPolicyAgentRuleDataSource(t *testing.T) {
 			{
 				Config: fmt.Sprintf(`
 				%s
-				data "datadog_csm_threats_agent_rules" "my_data_source" {
+				data "datadog_csm_threats_multi_policy_agent_rules" "my_data_source" {
 					policy_id = datadog_csm_threats_policy.policy_for_test.id
 				}
 				`, agentRuleConfig),
@@ -60,6 +62,30 @@ func TestAccCSMThreatsMultiPolicyAgentRuleDataSource(t *testing.T) {
 			},
 		},
 	})
+}
+
+func testAccCheckDatadogCSMThreatsMultiPolicyAgentRulesDataSourceConfig(policyName, agentRuleName string) string {
+	return fmt.Sprintf(`
+data "datadog_csm_threats_multi_policy_agent_rules" "my_data_source" {
+  policy_id = datadog_csm_threats_policy.policy.id
+}
+
+resource "datadog_csm_threats_policy" "policy" {
+  name = "%s"
+  enabled = true
+  description = "Test description"
+  tags = ["host_name:test_host"]
+}
+
+resource "datadog_csm_threats_multi_policy_agent_rule" "agent_rule" {
+  name = "%s"
+  description = "Test description"
+  enabled = true
+  expression = "open.file.name == \"etc/shadow/password\""
+  policy_id = datadog_csm_threats_policy.policy.id
+  product_tags = ["compliance_framework:PCI-DSS"]
+}
+`, policyName, agentRuleName)
 }
 
 func checkCSMThreatsMultiPolicyAgentRulesDataSourceContent(accProvider *fwprovider.FrameworkProvider, dataSourceName string, agentRuleName string) resource.TestCheckFunc {
@@ -120,6 +146,8 @@ func checkCSMThreatsMultiPolicyAgentRulesDataSourceContent(accProvider *fwprovid
 			resource.TestCheckResourceAttr(dataSourceName, fmt.Sprintf("agent_rules.%d.enabled", idx), "true"),
 			resource.TestCheckResourceAttr(dataSourceName, fmt.Sprintf("agent_rules.%d.description", idx), "im a rule"),
 			resource.TestCheckResourceAttr(dataSourceName, fmt.Sprintf("agent_rules.%d.expression", idx), "open.file.name == \"etc/shadow/password\""),
+			resource.TestCheckResourceAttr(dataSourceName, fmt.Sprintf("agent_rules.%d.product_tags.#", idx), "1"),
+			resource.TestCheckTypeSetElemAttr(dataSourceName, fmt.Sprintf("agent_rules.%d.product_tags.*", idx), "compliance_framework:PCI-DSS"),
 		)(state)
 	}
 }
